server_django/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
mongoengine==0.27.0
PyJWT==2.8.0
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
